ASM-4913 ensure rpm gem updates used by asm-deployer
The asm-deployer Gemfile has to depend on the git version of the
dell-asm-util gem so that Travis CI tests can run with the latest
version of that code. However, the appliance needs to use the version
of dell-asm-util that is included in the appliance Dell-ASM-Gems
rpm. Modify the asm-deployer rpm gradle build to simply remove the
:git tag from the dell-asm-util dependency so that the installed
dell-asm-util gem will be used.