ASM-8606 Run "bundle update --local" from post-install script
This used to be done in Dell-ASM-Gems, but it is more approriate
here. This rpm already has a dependency on Dell-ASM-Gems so that it
gets upgraded first. This way "bundle update --local" will be run
after the new gems are installed and after the asm-deployer Gemfile
has been updated to match.

Also removed the sed of Gemfile.lock since "bundle update --local"
will do it.