[misc] Add Gemfile.lock for asm-deployer
Previously the Gemfile.lock was not included in the asm-deployer
rpm. "bundle update --local" was run as a postinstall script in order
to generate it based on what gems were installed on the
appliance. That makes it very difficult to determine what gems are
_supposed_ to be included on the appliance.

This commit adds the Gemfile.lock to the rpm and removes the "bundle
update --local" post-install. Also tightens up permissions on
/opt/asm-deployer since we no longer expect torquebox to write into
that directory.