[maint] RPM fixes in preparation for CentOS 7
In general all Redline `Builder.addFile(...)` or
`Builder.addDirectory(...)` calls need to set the `addParents` boolean
to false. It unhelpfully defaults to true. When set to false it claims
ownership of all parent directories in addition to the actual path
specified. Yum in CentOS 7 will refuse to install RPMs that claim to
own the same directory as another RPM.